fixed warnings via cast

+++ serial_unit_prop_threaded/sudoku_solver_serial_unit_prop_threaded.cpp
@@ -224,7 +224,7 @@ int main(int argc, char *argv[]) {
       }
       toSolve.push_back(forSolving);
 
-      printf("toSolve size: %i\n", toSolve.size());
+      printf("toSolve size: %i\n", (int)toSolve.size());
 
       if (reverseSlot(p)) {
         currentDepth--;
@@ -272,7 +272,7 @@ int main(int argc, char *argv[]) {
   delete p;
   p = NULL;
 
-  printf("number of branches: %i\n", toSolve.size());
+  printf("number of branches: %i\n", (int)toSolve.size());
 
   p = solveInitializedPuzzles(&toSolve, highestVisitedPosition);
